// 04-countryApp/src/app/countries/interfaces/cache-store.interface.ts
import { Region } from "../enums/region";
import { Country } from "./country"

export interface CacheStore {
  byCapital: TermCountries;
  byCountries: TermCountries;
  byRegion: RegionCountries;
}

export interface TermCountries {
  term: string;
  countries: Country[];
}

export interface RegionCountries {
  region?: any;
  countries: Country[];
}